# Replace pending request channel with atomic integer.