<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

Other than the "active" queue, the pending queue never blocks the caller. Requests are either admitted or denied immmediately. We don't need to resort to channel ops for that and can instead use atomic integer operations for that.

## Reserve benchmark

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkBreakerReserve/sequential-16     123           93.3          -24.15%
BenchmarkBreakerReserve/parallel-16       465           395           -15.05%
```

## Maybe benchmark

```
benchmark                                    old ns/op     new ns/op     delta
BenchmarkBreakerMaybe/1-sequential-16        113           95.8          -15.22%
BenchmarkBreakerMaybe/1-parallel-16          398           309           -22.36%
BenchmarkBreakerMaybe/10-sequential-16       114           97.0          -14.91%
BenchmarkBreakerMaybe/10-parallel-16         1014          929           -8.38%
BenchmarkBreakerMaybe/100-sequential-16      111           96.6          -12.97%
BenchmarkBreakerMaybe/100-parallel-16        327           225           -31.19%
BenchmarkBreakerMaybe/1000-sequential-16     111           97.2          -12.43%
BenchmarkBreakerMaybe/1000-parallel-16       340           180           -47.06%
```

/assign @vagababov @julz