@markusthoemmes 2020-05-27
/test pull-knative-serving-autotls-tests