@vagababov 2020-05-27
After discussion in slack we concluded that it's totally fine :)